refactored types sub package
