+++ limit_order_book/order.py
@@ -1,9 +1,9 @@
 
-from limit_order_book.types.order_id import OrderId
-from limit_order_book.ticker import Ticker
-from limit_order_book.types.int_price import IntPrice
-from limit_order_book.types.volume import Volume
-from limit_order_book.order_side import OrderSide
+from limit_order_book.types import OrderId
+from limit_order_book.types import Ticker
+from limit_order_book.types import IntPrice
+from limit_order_book.types import Volume
+from limit_order_book.types import OrderSide
 from limit_order_book.trade import Trade
 
 from typeguard import typechecked
