+++ limit_order_book_webserver/convert_ticker_list_to_fastapi_ticker_list.py
@@ -1,5 +1,5 @@
 
-from limit_order_book.ticker import Ticker
+from limit_order_book.types import Ticker
 from limit_order_book_webserver.types import FastAPI_Ticker
