+++ tests/test_order_id.py
@@ -1,7 +1,7 @@
 
 import pytest
 
-from limit_order_book.types.order_id import OrderId
+from limit_order_book.types import OrderId
 
 
 def test_order_id_equal():
