+++ tests/test_single_side_limit_order_book/test_single_side_limit_order_book_find_order_by_order_id.py
@@ -1,10 +1,10 @@
 
-from limit_order_book.types.order_id import OrderId
-from limit_order_book.ticker import Ticker
-from limit_order_book.order_side import OrderSide
+from limit_order_book.types import OrderId
+from limit_order_book.types import Ticker
+from limit_order_book.types import OrderSide
 from limit_order_book.order import Order
-from limit_order_book.types.int_price import IntPrice
-from limit_order_book.types.volume import Volume
+from limit_order_book.types import IntPrice
+from limit_order_book.types import Volume
 
 from limit_order_book.single_side_limit_order_book import SingleSideLimitOrderBook
