+++ tests/test_webserver/test_webserver_api_top_of_book.py
@@ -4,7 +4,7 @@ from fastapi import status
 from fastapi.testclient import TestClient
 
 from limit_order_book_webserver.fastapi_webserver import app
-from limit_order_book.types.order_id import OrderId
+from limit_order_book.types import OrderId
 from limit_order_book.exceptions import DuplicateOrderIdError
 
 from tests.test_webserver.helper import helper_generate_order_without_order_id
